terraform/face_cut/requirements.txt
boto3
pillow
ydb